Many bug fixes and upgraded; new version
Implemented several changes to cope with getting Ultraspec under control.
In particular, the pipeline now translates the output-dependent coordinates
used to control Ultraspec (pixel 1 always = first to be read out) into output
independent coordinates so that the same format is the same part of the
chip. Also several improvements to rtplot / grab etc such as behaving properly
when no xml is found and when ctrl-C is hit. Other bug fixes are to shifter
and folder for a change in the headers of ucm. New programs are mostly python
scripts to do with bringing the logs under control; this is still work in
progress.