Changed statistics method to use subs code
The statistics method function in ultracam had a lot of duplicate
code which is in sigma_reject.cc within subs. This commit removes
this duplicate code and makes calls to the sigma_reject function
instead. Note that this commit changes the arguments to the
statisitics function; namely there is an addition 'careful' bool,
since sigma_reject in subs allows for an optional clipped mean
method whereby only a single pixel is rejected at a time, and also
the default argument for compute_median has been removed. Note that
the standard 'cavalier' method of sigma rejection is used as
standard within the code.