Fixes for timing issue of April-June 2010 NTT run
Owing to change of GPS, time stamps started to get taken after instead of before readouts. This is identical to a problem that occurred some time ago due to a change of clockboard and has the same fix. Thus various routines needed changing mainly the addition of a new date for chaning, and a small philosophical change such that I no longer think in terms of correct and incorrect in terms of the time stamps.